add referenceManager (#30)
* add referenceManager

* add project to sln

* separate logic from UI

* fix css loading issue, set default page, and prevent crash on missing integrator name

* fix reference list depth, fix display image

* improve reference rendering

* improve CSS for image viewer

Size to fit and hide the "tools" panel by default

* fix scene load path, show full path

* polish

* refactor common Integrator fields

* refactor reference image handling

* polish integrator selector

no need to track more than one integrator in this case

* reduce code duplication and upgrade to net10.0

* fix DocumentationReader namespace

* refactor scene file and reference management

ReferenceCache.cs now contains all logic to manage references,
SceneConfig only specifies parameters (like min and max depth) to be
used for a scene, and SceneRegistry only finds folders but does not
contain any additional logic. Scene loading and blender importing is
moved to SceneLoader. This also allows better auto-import because every
time a scene is instanced, we can now easily check for blender file
changes.

* update dependencies

* update simpleimageio

* update templates

* reorganize, separate benchmarking and scene management

* update todo notes

* update reference manager after refactor

* implement AddSamples logic

* fixes

* more fixes

* polish

* fix warning

---------

+++ MaterialTest/Pages/IntegratorTest.razor.cs
@@ -1,5 +1,4 @@
 using Microsoft.AspNetCore.Components;
-using SeeSharp.Blazor;
 
 namespace MaterialTest.Pages;
 
@@ -22,7 +21,7 @@ public partial class IntegratorTest : ComponentBase
         scene.Prepare();
         VertexConnectionAndMerging vcm = new()
         {
-            NumIterations = NumSamples,
+            NumIterations = (uint)NumSamples,
             MaxDepth = MaxDepth,
             RenderTechniquePyramid = true
         };
